extended file conversion functionality (still work in progress)
reworked the nestedStringConverter class to be more generally applicable.
seperating a given string is now handled by the class itself, only the conversion from a list of strings to an object has to be implemented in the type-specific implementations.
Also added a sanitize()-Operation that removes all reserved characters from a given string